Added Dog and popping and creating bublbles
Moved in all the images that were made so far. Imported the dog image (other may still need to be scaled). Made the Bubble into a prefab. Made it so the dog can make a bubble by pressing "jump" when there is no bubble. Made the bubble pop when you press control or east button